refactor(lb): added load_balancing , upstream_pool, hash modules and updated: backend_pool module scope
