+++ crates/lb/src/lib.rs
@@ -1,3 +1,10 @@
+pub mod backend;
+pub mod backend_pool;
+pub mod hash;
+pub mod health;
+pub mod load_balancing;
+pub mod upstream_pool;
+
 use std::{
     cell::RefCell,
     sync::{
@@ -12,559 +19,6 @@ use spooky_config::config::{Backend, HealthCheck};
 
 thread_local! {
     static LB_RANDOM_RNG: RefCell<StdRng> = RefCell::new(StdRng::from_entropy());
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum HealthFailureReason {
-    HttpStatus5xx,
-    Timeout,
-    Transport,
-    Tls,
-    CircuitOpen,
-}
-
-const DEFAULT_REPLICAS: u32 = 64;
-const FNV_OFFSET: u64 = 0xcbf29ce484222325;
-const FNV_PRIME: u64 = 0x00000100000001b3;
-
-#[derive(Clone)]
-pub struct BackendState {
-    address: String,
-    weight: u32,
-    health_check: Option<HealthCheck>,
-    consecutive_failures: u32,
-    health_state: HealthState,
-    active_requests: Arc<AtomicUsize>,
-    ewma_latency_ms: Option<f64>,
-}
-
-#[derive(Clone)]
-enum HealthState {
-    Healthy,
-    // `reason` is stored for future introspection; suppressed until wired to metrics
-    #[allow(dead_code)]
-    Unhealthy {
-        until: Instant,
-        successes: u32,
-        reason: HealthFailureReason,
-    },
-}
-
-pub enum HealthTransition {
-    BecameHealthy,
-    BecameUnhealthy,
-}
-
-impl BackendState {
-    pub fn new(backend: &Backend) -> Self {
-        Self {
-            address: backend.address.clone(),
-            weight: backend.weight.max(1),
-            health_check: backend.health_check.clone(),
-            consecutive_failures: 0,
-            health_state: HealthState::Healthy,
-            active_requests: Arc::new(AtomicUsize::new(0)),
-            ewma_latency_ms: None,
-        }
-    }
-
-    pub fn is_healthy(&self) -> bool {
-        matches!(self.health_state, HealthState::Healthy)
-    }
-
-    /// Returns true when an active health-check loop is running for this backend.
-    /// When active checks are present, only the health-check loop should drive
-    /// consecutive_failures — request-path failures should not contribute.
-    pub fn has_active_health_check(&self) -> bool {
-        self.health_check.as_ref().is_some_and(|hc| hc.interval > 0)
-    }
-
-    pub fn address(&self) -> &str {
-        &self.address
-    }
-
-    pub fn health_check(&self) -> Option<&HealthCheck> {
-        self.health_check.as_ref()
-    }
-
-    pub fn weight(&self) -> u32 {
-        self.weight
-    }
-
-    pub fn active_requests(&self) -> usize {
-        self.active_requests.load(Ordering::Relaxed)
-    }
-
-    pub fn ewma_latency_ms(&self) -> Option<f64> {
-        self.ewma_latency_ms
-    }
-
-    pub fn record_success(&mut self) -> Option<HealthTransition> {
-        match &mut self.health_state {
-            HealthState::Healthy => {
-                self.consecutive_failures = 0;
-                None
-            }
-            HealthState::Unhealthy {
-                until, successes, ..
-            } => {
-                if Instant::now() < *until {
-                    return None;
-                }
-
-                *successes += 1;
-                let success_threshold = self
-                    .health_check
-                    .as_ref()
-                    .map_or(1, |hc| hc.success_threshold);
-                if *successes >= success_threshold {
-                    self.consecutive_failures = 0;
-                    self.health_state = HealthState::Healthy;
-                    return Some(HealthTransition::BecameHealthy);
-                }
-                None
-            }
-        }
-    }
-
-    pub fn record_failure(&mut self, reason: HealthFailureReason) -> Option<HealthTransition> {
-        self.consecutive_failures = self.consecutive_failures.saturating_add(1);
-        let threshold = self
-            .health_check
-            .as_ref()
-            .map_or(3, |hc| hc.failure_threshold);
-        if self.consecutive_failures < threshold {
-            return None;
-        }
-
-        self.consecutive_failures = 0;
-        let cooldown = Duration::from_millis(
-            self.health_check
-                .as_ref()
-                .map_or(10_000, |hc| hc.cooldown_ms),
-        );
-        self.health_state = HealthState::Unhealthy {
-            until: Instant::now() + cooldown,
-            successes: 0,
-            reason,
-        };
-        Some(HealthTransition::BecameUnhealthy)
-    }
-
-    /// Cooldown expiry, if this backend is currently unhealthy.
-    fn cooldown_until(&self) -> Option<Instant> {
-        if let HealthState::Unhealthy { until, .. } = self.health_state {
-            Some(until)
-        } else {
-            None
-        }
-    }
-
-    /// Optimistically re-admit an ejected backend once its cooldown has elapsed
-    /// so live traffic can probe it again. Returns true on transition.
-    fn readmit_if_expired(&mut self, now: Instant) -> bool {
-        if let HealthState::Unhealthy { until, .. } = self.health_state
-            && now >= until
-        {
-            self.consecutive_failures = 0;
-            self.health_state = HealthState::Healthy;
-            return true;
-        }
-        false
-    }
-}
-
-pub struct BackendPool {
-    backends: Vec<BackendState>,
-    healthy: Vec<usize>,
-    healthy_pos: Vec<Option<usize>>,
-    membership_epoch: u64,
-    // Earliest cooldown expiry among passively-ejected backends (no active
-    // health check), driving time-based re-admission. `None` when none pending.
-    earliest_readmit: Option<Instant>,
-}
-
-pub struct UpstreamPool {
-    pub pool: BackendPool,
-    pub load_balancer: LoadBalancing,
-    lb_key: Option<String>,
-}
-
-impl UpstreamPool {
-    pub fn from_upstream(upstream: &spooky_config::config::Upstream) -> Result<Self, String> {
-        let backends = upstream.backends.iter().map(BackendState::new).collect();
-
-        let load_balancer = LoadBalancing::from_config(&upstream.load_balancing.lb_type)?;
-        let lb_key = upstream
-            .load_balancing
-            .key
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_string);
-
-        Ok(Self {
-            pool: BackendPool::new_from_states(backends),
-            load_balancer,
-            lb_key,
-        })
-    }
-
-    pub fn pick(&mut self, key: &str) -> Option<usize> {
-        self.pool.reconcile_readmit();
-        let selected = self.load_balancer.pick(key, &self.pool)?;
-        self.pool.begin_request(selected);
-        Some(selected)
-    }
-
-    pub fn pick_readonly(&self, key: &str) -> Option<usize> {
-        self.load_balancer.pick_readonly(key, &self.pool)
-    }
-
-    pub fn pick_without_begin(&mut self, key: &str) -> Option<usize> {
-        self.pool.reconcile_readmit();
-        self.load_balancer.pick(key, &self.pool)
-    }
-
-    pub fn begin_request_if_healthy(&self, index: usize) -> bool {
-        if self.pool.is_healthy_index(index) {
-            self.pool.begin_request(index);
-            true
-        } else {
-            false
-        }
-    }
-
-    pub fn finish_request(&mut self, index: usize, latency: Duration, status: Option<u16>) {
-        self.pool.finish_request(index, latency, status);
-    }
-
-    pub fn lb_name(&self) -> &'static str {
-        self.load_balancer.name()
-    }
-
-    pub fn lb_key(&self) -> Option<&str> {
-        self.lb_key.as_deref()
-    }
-}
-
-impl BackendPool {
-    pub fn new_from_states(backends: Vec<BackendState>) -> Self {
-        let mut healthy = Vec::with_capacity(backends.len());
-        let mut healthy_pos = vec![None; backends.len()];
-
-        for (idx, backend) in backends.iter().enumerate() {
-            if backend.is_healthy() {
-                healthy_pos[idx] = Some(healthy.len());
-                healthy.push(idx);
-            }
-        }
-
-        Self {
-            backends,
-            healthy,
-            healthy_pos,
-            membership_epoch: 0,
-            earliest_readmit: None,
-        }
-    }
-
-    pub fn len(&self) -> usize {
-        self.backends.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.backends.is_empty()
-    }
-
-    pub fn address(&self, index: usize) -> Option<&str> {
-        self.backends.get(index).map(|b| b.address())
-    }
-
-    pub fn mark_success(&mut self, index: usize) -> Option<HealthTransition> {
-        if index >= self.backends.len() {
-            return None;
-        }
-
-        let (was_healthy, is_healthy, transition) = {
-            let backend = &mut self.backends[index];
-            let was_healthy = backend.is_healthy();
-            let transition = backend.record_success();
-            let is_healthy = backend.is_healthy();
-            (was_healthy, is_healthy, transition)
-        };
-
-        if was_healthy != is_healthy {
-            if is_healthy {
-                debug_assert!(self.mark_healthy(index));
-            } else {
-                debug_assert!(self.mark_unhealthy(index));
-            }
-            self.membership_epoch = self.membership_epoch.wrapping_add(1);
-        }
-
-        transition
-    }
-
-    /// Mark a failure from the active health-check loop — always recorded.
-    pub fn mark_failure(&mut self, index: usize) -> Option<HealthTransition> {
-        self.mark_failure_with_reason(index, HealthFailureReason::HttpStatus5xx)
-    }
-
-    /// Mark a failure from the request path (passive).
-    /// Skipped when an active health-check loop is running for this backend,
-    /// because the loop is the sole authority on consecutive_failures in that case.
-    pub fn mark_request_failure(
-        &mut self,
-        index: usize,
-        reason: HealthFailureReason,
-    ) -> Option<HealthTransition> {
-        if index < self.backends.len() && self.backends[index].has_active_health_check() {
-            return None;
-        }
-        self.mark_failure_with_reason(index, reason)
-    }
-
-    pub fn mark_failure_with_reason(
-        &mut self,
-        index: usize,
-        reason: HealthFailureReason,
-    ) -> Option<HealthTransition> {
-        if index >= self.backends.len() {
-            return None;
-        }
-
-        let (was_healthy, is_healthy, transition) = {
-            let backend = &mut self.backends[index];
-            let was_healthy = backend.is_healthy();
-            let transition = backend.record_failure(reason);
-            let is_healthy = backend.is_healthy();
-            (was_healthy, is_healthy, transition)
-        };
-
-        if was_healthy != is_healthy {
-            if is_healthy {
-                debug_assert!(self.mark_healthy(index));
-            } else {
-                debug_assert!(self.mark_unhealthy(index));
-                // Passive ejections have no active loop to recover them; record
-                // the cooldown so `reconcile_readmit` can re-admit on expiry.
-                if !self.backends[index].has_active_health_check()
-                    && let Some(until) = self.backends[index].cooldown_until()
-                {
-                    self.earliest_readmit =
-                        Some(self.earliest_readmit.map_or(until, |e| e.min(until)));
-                }
-            }
-            self.membership_epoch = self.membership_epoch.wrapping_add(1);
-        }
-
-        transition
-    }
-
-    /// True when any backend is passively ejected and pending re-admission.
-    /// Clock-free so the read-locked hot path pays only a branch (no syscall):
-    /// while something is pending, callers take the write-locked slow path where
-    /// `reconcile_readmit` checks the actual cooldown clock.
-    pub fn readmit_due(&self) -> bool {
-        self.earliest_readmit.is_some()
-    }
-
-    /// Re-admit passively-ejected backends whose cooldown has elapsed so live
-    /// traffic can probe them. Reads the clock only when a re-admission is
-    /// actually pending, keeping the healthy pick path syscall-free.
-    pub fn reconcile_readmit(&mut self) {
-        if self.earliest_readmit.is_some() {
-            self.reconcile_readmit_at(Instant::now());
-        }
-    }
-
-    /// Core of [`reconcile_readmit`] with an injectable clock. Recomputes the
-    /// next pending expiry; early-returns while the soonest cooldown is unmet.
-    fn reconcile_readmit_at(&mut self, now: Instant) {
-        let Some(earliest) = self.earliest_readmit else {
-            return;
-        };
-        if now < earliest {
-            return;
-        }
-        let mut next: Option<Instant> = None;
-        for index in 0..self.backends.len() {
-            if self.backends[index].has_active_health_check() {
-                continue;
-            }
-            if self.backends[index].readmit_if_expired(now) {
-                debug_assert!(self.mark_healthy(index));
-                self.membership_epoch = self.membership_epoch.wrapping_add(1);
-            } else if let Some(until) = self.backends[index].cooldown_until() {
-                next = Some(next.map_or(until, |e| e.min(until)));
-            }
-        }
-        self.earliest_readmit = next;
-    }
-
-    pub fn health_check(&self, index: usize) -> Option<HealthCheck> {
-        self.backends
-            .get(index)
-            .and_then(|b| b.health_check().cloned())
-    }
-
-    pub fn healthy_indices(&self) -> Vec<usize> {
-        self.healthy.clone()
-    }
-
-    pub fn healthy_len(&self) -> usize {
-        self.healthy.len()
-    }
-
-    pub fn healthy_indices_iter(&self) -> impl Iterator<Item = usize> + '_ {
-        self.healthy.iter().copied()
-    }
-
-    pub fn all_indices(&self) -> Vec<usize> {
-        (0..self.backends.len()).collect()
-    }
-
-    pub fn backend(&self, index: usize) -> Option<&BackendState> {
-        self.backends.get(index)
-    }
-
-    pub fn membership_epoch(&self) -> u64 {
-        self.membership_epoch
-    }
-
-    pub fn is_healthy_index(&self, index: usize) -> bool {
-        self.healthy_pos.get(index).copied().flatten().is_some()
-    }
-
-    pub fn begin_request(&self, index: usize) {
-        if let Some(backend) = self.backends.get(index) {
-            backend.active_requests.fetch_add(1, Ordering::Relaxed);
-        }
-    }
-
-    pub fn finish_request(&mut self, index: usize, latency: Duration, status: Option<u16>) {
-        let Some(backend) = self.backends.get_mut(index) else {
-            return;
-        };
-
-        let _ =
-            backend
-                .active_requests
-                .fetch_update(Ordering::Relaxed, Ordering::Relaxed, |current| {
-                    Some(current.saturating_sub(1))
-                });
-
-        if status.is_some_and(|code| (500..=599).contains(&code)) {
-            return;
-        }
-
-        let observed_ms = latency.as_secs_f64() * 1_000.0;
-        let alpha = 0.2_f64;
-        backend.ewma_latency_ms = Some(match backend.ewma_latency_ms {
-            Some(previous) => alpha * observed_ms + (1.0 - alpha) * previous,
-            None => observed_ms,
-        });
-    }
-
-    fn mark_healthy(&mut self, index: usize) -> bool {
-        if index >= self.backends.len() {
-            return false;
-        }
-
-        if self.healthy_pos[index].is_some() {
-            return false;
-        }
-
-        let pos = self.healthy.len();
-        self.healthy.push(index);
-        self.healthy_pos[index] = Some(pos);
-        true
-    }
-
-    fn mark_unhealthy(&mut self, index: usize) -> bool {
-        if index >= self.backends.len() {
-            return false;
-        }
-
-        let Some(pos) = self.healthy_pos[index] else {
-            return false;
-        };
-
-        let removed = self.healthy.swap_remove(pos);
-        debug_assert_eq!(removed, index);
-
-        if pos < self.healthy.len() {
-            let moved_index = self.healthy[pos];
-            self.healthy_pos[moved_index] = Some(pos);
-        }
-
-        self.healthy_pos[index] = None;
-        true
-    }
-}
-
-pub enum LoadBalancing {
-    RoundRobin(RoundRobin),
-    ConsistentHash(ConsistentHash),
-    Random(Random),
-    LeastConnections(LeastConnections),
-    LatencyAware(LatencyAware),
-    StickyCid(StickyCid),
-}
-
-impl LoadBalancing {
-    pub fn from_config(value: &str) -> Result<Self, String> {
-        let mode = value.trim().to_lowercase();
-        match mode.as_str() {
-            "round-robin" | "round_robin" | "rr" => Ok(Self::RoundRobin(RoundRobin::new())),
-            "consistent-hash" | "consistent_hash" | "ch" => {
-                Ok(Self::ConsistentHash(ConsistentHash::new(DEFAULT_REPLICAS)))
-            }
-            "random" => Ok(Self::Random(Random::new())),
-            "least-connections" | "least_connections" | "lc" => {
-                Ok(Self::LeastConnections(LeastConnections::new()))
-            }
-            "latency-aware" | "latency_aware" | "la" => Ok(Self::LatencyAware(LatencyAware::new())),
-            "sticky-cid" | "sticky_cid" | "cid-sticky" | "cid_sticky" => {
-                Ok(Self::StickyCid(StickyCid::new(DEFAULT_REPLICAS)))
-            }
-            _ => Err(format!("unsupported load balancing type: {value}")),
-        }
-    }
-
-    pub fn pick(&mut self, key: &str, pool: &BackendPool) -> Option<usize> {
-        match self {
-            LoadBalancing::RoundRobin(rr) => rr.pick(pool),
-            LoadBalancing::ConsistentHash(ch) => ch.pick(key, pool),
-            LoadBalancing::Random(rand) => rand.pick(pool),
-            LoadBalancing::LeastConnections(lc) => lc.pick(pool),
-            LoadBalancing::LatencyAware(la) => la.pick(pool),
-            LoadBalancing::StickyCid(sticky) => sticky.pick(key, pool),
-        }
-    }
-
-    pub fn pick_readonly(&self, _key: &str, pool: &BackendPool) -> Option<usize> {
-        match self {
-            LoadBalancing::RoundRobin(rr) => rr.pick_readonly(pool),
-            LoadBalancing::Random(rand) => rand.pick_readonly(pool),
-            LoadBalancing::LeastConnections(lc) => lc.pick_readonly(pool),
-            LoadBalancing::LatencyAware(la) => la.pick_readonly(pool),
-            // ConsistentHash and StickyCid keep mutable ring caches.
-            LoadBalancing::ConsistentHash(_) | LoadBalancing::StickyCid(_) => None,
-        }
-    }
-
-    pub fn name(&self) -> &'static str {
-        match self {
-            LoadBalancing::RoundRobin(_) => "round-robin",
-            LoadBalancing::ConsistentHash(_) => "consistent-hash",
-            LoadBalancing::Random(_) => "random",
-            LoadBalancing::LeastConnections(_) => "least-connections",
-            LoadBalancing::LatencyAware(_) => "latency-aware",
-            LoadBalancing::StickyCid(_) => "sticky-cid",
-        }
-    }
 }
 
 pub struct RoundRobin {
@@ -810,51 +264,3 @@ impl StickyCid {
         self.inner.pick(key, pool)
     }
 }
-
-fn expected_ring_entries(pool: &BackendPool, replicas: u32) -> usize {
-    pool.healthy
-        .iter()
-        .map(|&idx| replicas.saturating_mul(pool.backends[idx].weight()) as usize)
-        .sum()
-}
-
-fn hash_backend_replica(address: &str, replica: u32) -> u64 {
-    let mut hash = FNV_OFFSET;
-    for &byte in address.as_bytes() {
-        hash = hash64_update(hash, byte);
-    }
-    hash = hash64_update(hash, b'-');
-
-    let mut digits = [0u8; 10];
-    let mut value = replica;
-    let mut cursor = digits.len();
-    loop {
-        cursor -= 1;
-        digits[cursor] = b'0' + (value % 10) as u8;
-        value /= 10;
-        if value == 0 {
-            break;
-        }
-    }
-
-    for &digit in &digits[cursor..] {
-        hash = hash64_update(hash, digit);
-    }
-
-    hash
-}
-
-fn hash64(data: &[u8]) -> u64 {
-    let mut hash = FNV_OFFSET;
-    for byte in data {
-        hash = hash64_update(hash, *byte);
-    }
-    hash
-}
-
-fn hash64_update(hash: u64, byte: u8) -> u64 {
-    (hash ^ byte as u64).wrapping_mul(FNV_PRIME)
-}
-
-#[cfg(test)]
-mod tests;
